fix(TextField): change type of component properties
- Made a prohibition on the use of properties minRows&maxRows / rows together

- Change description of props: rows, minRows, maxRows

issue -  #976
