+++ src/components/Badge/__stories__/examples/BadgeExampleForm/BadgeExampleForm.tsx
@@ -6,7 +6,7 @@ import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Badge } from '../../../Badge';
 
 export const BadgeExampleForm = () => (
-  <div className={cnDocsDecorator('Section', [cnDocsExample(null)])}>
+  <div className={cnDocsDecorator('Section', [cnDocsExample()])}>
     <Badge form="default" label="Default badge" className={wp.decorator({ 'indent-r': 'm' })} />
     <Badge form="round" label="Round badge" className={wp.decorator({ 'indent-r': 'm' })} />
   </div>
