+++ src/components/Badge/__stories__/examples/BadgeExampleMinified/BadgeExampleMinified.tsx
@@ -6,7 +6,7 @@ import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Badge } from '../../../Badge';
 
 export const BadgeExampleMinified = () => (
-  <div className={cnDocsDecorator('Section', [cnDocsExample(null)])}>
+  <div className={cnDocsDecorator('Section', [cnDocsExample()])}>
     <Badge minified status="error" label="Стойте" className={wp.decorator({ 'indent-r': 'm' })} />
     <Badge minified status="success" label="Идите" className={wp.decorator({ 'indent-r': 'm' })} />
   </div>
