+++ src/components/Badge/__stories__/examples/BadgeExampleSize/BadgeExampleSize.tsx
@@ -6,7 +6,7 @@ import * as wp from '../../../../../uiKit/whitepaper/whitepaper';
 import { Badge } from '../../../Badge';
 
 export const BadgeExampleSize = () => (
-  <div className={cnDocsDecorator('Section', [cnDocsExample(null)])}>
+  <div className={cnDocsDecorator('Section', [cnDocsExample()])}>
     <Badge size="s" label="Badge s" className={wp.decorator({ 'indent-r': 'm' })} />
     <Badge size="m" label="Badge m" className={wp.decorator({ 'indent-r': 'm' })} />
     <Badge size="l" label="Badge l" className={wp.decorator({ 'indent-r': 'm' })} />
