+++ src/components/Badge/__stories__/examples/BadgeExampleView/BadgeExampleView.tsx
@@ -7,10 +7,7 @@ import { Badge } from '../../../Badge';
 
 export const BadgeExampleView = () => (
   <div
-    className={cnDocsDecorator('Section', [
-      cnDocsExample(null),
-      wp.decorator({ 'indent-t': 'xl' }),
-    ])}
+    className={cnDocsDecorator('Section', [cnDocsExample(), wp.decorator({ 'indent-t': 'xl' })])}
   >
     <Badge view="filled" label="Filled badge" className={wp.decorator({ 'indent-r': 'm' })} />
     <Badge view="stroked" label="Stroked badge" className={wp.decorator({ 'indent-r': 'm' })} />
